создание рейтинга пользователя и общего рейтинга туров создание комментариев

+++ console/migrations/m220222_130049_add_column_rating_to_list_resorts.php
@@ -9,7 +9,7 @@ class m220222_130049_add_column_rating_to_list_resorts extends Migration
 {
     public function up()
     {
-        $this->addColumn('{{list_resorts}}', 'rating', $this->integer()->defaultValue(0));
+        $this->addColumn('{{list_resorts}}', 'rating', $this->float()->defaultValue(0));
     }
 
     /**
